Actualización del firmware para leer NPK

+++ firmware/include/hardware.h
@@ -8,4 +8,4 @@ void setupHardware();
 void processHumidity_Temp(int &humedad, int &temp);
 
 // Pasa las variables por referencia para actualizar los 3 valores de NPK
-void readNPKValues(byte &n, byte &p, byte &k);
+void readNPKValues(uint16_t &n, uint16_t &p, uint16_t &k);
